from local

+++ fairseq/criterions/ctc.py
@@ -3053,7 +3053,7 @@ class L2SCriterion(FairseqCriterion):
             with torch.no_grad():
                 lm_output = self.lm(**lm_input)
                 lm_output = lm_output['last_hidden_state']
-                #lm_output = lm_output.half()
+                lm_output = lm_output.half()
             
             am_output = net_output['encoder_feat'].transpose(0, 1) ## T x B x C -> B x T x C
             am_output = self.lm_decoder(am_output)
